refactor: useSearchingMode hook 추가
기존의 useMode 훅을 리팩터링함.
기존 search/trending 문자열을 기준으로 렌더링하던 것에서 isSearching이 true/false에 따라 렌더링하는 것으로 변경
