+++ src/pages/Search/components/RelatedSearches/index.tsx
@@ -1,20 +1,20 @@
 import cn from 'utils/ts/classNames';
 import styles from 'pages/Search/components/RelatedSearches/RelatedSearches.module.scss';
-import MODE from 'pages/Search/static/mode';
 import suggestion from 'pages/Search/static/suggestion';
+import useSearchingMode from 'pages/Search/hooks/useSearchingMode';
 import RelatedItem from './components/RelatedItem';
 
 interface Props {
-  mode: string,
   text: string,
 }
 
-export default function RelatedSearches({ mode, text }: Props) {
+export default function RelatedSearches({ text }: Props) {
+  const isSearching = useSearchingMode();
   return (
     <div className={styles.search}>
       <ul className={cn({
         [styles['search-related-list']]: true,
-        [styles['search-related-list--hidden']]: mode === MODE.trending,
+        [styles['search-related-list--hidden']]: !isSearching,
       })}
       >
         {text === '' ? null : suggestion.filter((item) => item.title.includes(text))
